친구 조회 출력, api 형식 수정
