+++ prisma/schema.prisma
@@ -31,7 +31,9 @@ model crops {
   id           String       @db.VarChar(100)
   user_id      String       @db.VarChar(100)
   type         crops_type
-  planted_at   DateTime     @db.Timestamp(6)
+  planted_at   DateTime     @db.Timestamptz(6)
+  harvested_at DateTime?    @db.Timestamptz(6)
+  grade        fruit_grade?
   user_account user_account @relation(fields: [user_id], references: [id], onDelete: NoAction, onUpdate: NoAction, map: "fk_user_account_to_crops_1")
 
   @@id([id, user_id], map: "pk_crops")
@@ -46,17 +48,6 @@ model friend {
   user_account_friend_requester_idTouser_account user_account @relation("friend_requester_idTouser_account", fields: [requester_id], references: [id], onDelete: NoAction, onUpdate: NoAction, map: "fk_friend_user_account_2")
 
   @@id([requester_id, acceptor_id], map: "pk_friend")
-}
-
-model harvested_crops {
-  id           String       @db.VarChar(100)
-  user_id      String       @db.VarChar(100)
-  type         crops_type
-  harvested_at DateTime     @db.Timestamp(6)
-  grade        crops_type
-  user_account user_account @relation(fields: [user_id], references: [id], onDelete: NoAction, onUpdate: NoAction, map: "fk_user_account_to_harvested_crops_1")
-
-  @@id([id, user_id], map: "pk_harvested_crops")
 }
 
 model organization {
@@ -74,6 +65,7 @@ model report {
   content                                       String          @db.VarChar(500)
   reported_at                                   DateTime        @db.Timestamp(6)
   screenshot                                    String?         @db.VarChar(500)
+  accepted                                      Boolean?
   user_account_report_reporter_idTouser_account user_account    @relation("report_reporter_idTouser_account", fields: [reporter_id], references: [id], onDelete: NoAction, onUpdate: NoAction, map: "report_reporter_id__fk")
   user_account_report_suspect_idTouser_account  user_account    @relation("report_suspect_idTouser_account", fields: [suspect_id], references: [id], onDelete: NoAction, onUpdate: NoAction, map: "report_suspect_id__fk")
 }
@@ -125,20 +117,19 @@ model tag {
 }
 
 model user_account {
-  id                                       String            @id(map: "pk_user_account") @db.VarChar(100)
-  name                                     String            @db.VarChar(20)
-  introduce                                String?           @db.VarChar(100)
+  id                                       String          @id(map: "pk_user_account") @db.VarChar(100)
+  name                                     String          @db.VarChar(20)
+  introduce                                String?         @db.VarChar(100)
   score                                    BigInt
-  profile_image                            String?           @db.VarChar(255)
+  profile_image                            String?         @db.VarChar(255)
   status                                   user_status
   belong                                   belong[]
   block                                    block[]
   crops                                    crops[]
-  friend_friend_acceptor_idTouser_account  friend[]          @relation("friend_acceptor_idTouser_account")
-  friend_friend_requester_idTouser_account friend[]          @relation("friend_requester_idTouser_account")
-  harvested_crops                          harvested_crops[]
-  report_report_reporter_idTouser_account  report[]          @relation("report_reporter_idTouser_account")
-  report_report_suspect_idTouser_account   report[]          @relation("report_suspect_idTouser_account")
+  friend_friend_acceptor_idTouser_account  friend[]        @relation("friend_acceptor_idTouser_account")
+  friend_friend_requester_idTouser_account friend[]        @relation("friend_requester_idTouser_account")
+  report_report_reporter_idTouser_account  report[]        @relation("report_reporter_idTouser_account")
+  report_report_suspect_idTouser_account   report[]        @relation("report_suspect_idTouser_account")
   room                                     room[]
   study_history                            study_history[]
   user_tag                                 user_tag[]
