+++ src/models/organization/BelongOrganization.ts
@@ -1,0 +1,5 @@
+export interface BelongOrganization {
+  readonly userId: string;
+  readonly organizationId: string;
+  readonly organizationName: string;
+}
