+++ src/models/organization/Organization.ts
@@ -1,0 +1,5 @@
+export interface Organization {
+  readonly id: string;
+  readonly name: string;
+  readonly address: string;
+}
