+++ src/models/organization/OrganizationVerifyEmailForm.ts
@@ -1,0 +1,7 @@
+export interface OrganizationVerifyEmailForm {
+  readonly userId: string;
+  readonly userName: string;
+  readonly email: string;
+  readonly organizationId: string;
+  readonly organizationName: string;
+}
